requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
Jinja2==3.1.4

